Schreiben in die Datenbank implementiert, geänderte Bezeichner

+++ forecast_accuracy/ForecastObject.cs
@@ -8,13 +8,16 @@ namespace forecast_accuracy
 {
     public class ForecastObject
     {
-        public int Dt { get; set; }
+        // Unix Timestamp
+        public long Dt { get; set; }
         public ForecastMain Main { get; set; }
         public List<Weather> Weather { get; set; }
         public Clouds Clouds { get; set; }
         public Wind Wind { get; set; }
-        public Rain Rain { get; set; } = new Rain(); // Kein Regen -> "default" Objekt mit wert 0
+        // Kein Regen -> "default" Objekt mit wert 0
+        public Rain Rain { get; set; } = new Rain();
         public ForecastSys Sys { get; set; } // Testen ob es ohne funktioniert
-        public DateTime Dt_text { get; set; } // In JSON als String im Format "YYYY-MM-DD hh:mm:ss"
+        // In JSON als String im Format "YYYY-MM-DD hh:mm:ss"
+        public DateTime Dt_text { get; set; }
     }
 }
